[ADD] support for multipage workspace
this commit adds the support for multipage workspace , user can now add infinite pages into the workspace each running on its own dataflow separated from other workspaces, even zoom level are designed to be configurable for each workspace page, user can now open infinite *.qbp files in the app
